refactor: cria função isAdmin e move feature toggle para utils
