+++ src/utils/isAdmin/index.js
@@ -1,0 +1,4 @@
+export const isAdmin = () => {
+  const role = localStorage.getItem('role')
+  return role === 'admin'
+}
